requirements.txt
transformers
datasets
pandas
evaluate
ipykernel
hf_xet
torch
nltk
matplotlib
seaborn
rouge_score
accelerate>=0.26.0
pyyaml
python-box==6.0.2